fix: fix lint errors

+++ contracts/interfaces/PythInterface.sol
@@ -62,7 +62,6 @@ contract PythStructs {
         HALTED,
         AUCTION
     }
-
 }
 
 /// @title Consume prices from the Pyth Network (https://pyth.network/).
@@ -72,26 +71,36 @@ interface IPyth {
     /// @param id The Pyth Price Feed ID.
     /// @param fresh True if the price update is more recent and stored.
     /// @param chainId ID of the source chain that the batch price update containing this price.
-    /// This value comes from Wormhole, and you can find the corresponding chains at https://docs.wormholenetwork.com/wormhole/contracts.
+    /// This value comes from Wormhole,
+    /// you can find the corresponding chains at https://docs.wormholenetwork.com/wormhole/contracts.
     /// @param sequenceNumber Sequence number of the batch price update containing this price.
     /// @param lastPublishTime Publish time of the previously stored price.
     /// @param publishTime Publish time of the given price update.
     /// @param price Current price of the given price update.
     /// @param conf Current confidence interval of the given price update.
-    event PriceFeedUpdate(bytes32 indexed id, bool indexed fresh, uint16 chainId, uint64 sequenceNumber, uint64 lastPublishTime, uint64 publishTime, int64 price, uint64 conf);
+    event PriceFeedUpdate(
+        bytes32 indexed id,
+        bool indexed fresh,
+        uint16 chainId,
+        uint64 sequenceNumber,
+        uint64 lastPublishTime,
+        uint64 publishTime,
+        int64 price,
+        uint64 conf
+    );
 
     /// @dev Emitted when a batch price update is processed successfully.
     /// @param chainId ID of the source chain that the batch price update comes from.
     /// @param sequenceNumber Sequence number of the batch price update.
     /// @param batchSize Number of prices within the batch price update.
     /// @param freshPricesInBatch Number of prices that were more recent and were stored.
-    event BatchPriceFeedUpdate(uint16 chainId, uint64 sequenceNumber, uint batchSize, uint freshPricesInBatch);
+    event BatchPriceFeedUpdate(uint16 chainId, uint64 sequenceNumber, uint256 batchSize, uint256 freshPricesInBatch);
 
     /// @dev Emitted when a call to `updatePriceFeeds` is processed successfully.
     /// @param sender Sender of the call (`msg.sender`).
     /// @param batchCount Number of batches that this function processed.
     /// @param fee Amount of paid fee for updating the prices.
-    event UpdatePriceFeeds(address indexed sender, uint batchCount, uint fee);
+    event UpdatePriceFeeds(address indexed sender, uint256 batchCount, uint256 fee);
 
     /// @notice Returns the current price and confidence interval.
     /// @dev Reverts if the current price is not available.
@@ -118,13 +127,19 @@ interface IPyth {
     /// safer / easier to use either `getCurrentPrice` or `getLatestAvailablePriceWithinDuration`.
     /// @return price - please read the documentation of PythStructs.Price to understand how to use this safely.
     /// @return publishTime - the UNIX timestamp of when this price was computed.
-    function getLatestAvailablePriceUnsafe(bytes32 id) external view returns (PythStructs.Price memory price, uint64 publishTime);
+    function getLatestAvailablePriceUnsafe(bytes32 id)
+        external
+        view
+        returns (PythStructs.Price memory price, uint64 publishTime);
 
     /// @notice Returns the latest price as long as it was updated within `duration` seconds of the current time.
     /// @dev This function is a sanity-checked version of `getLatestAvailablePriceUnchecked` which is useful in
     /// applications that require a sufficiently-recent price. Reverts if the price wasn't updated sufficiently
     /// recently.
-    function getLatestAvailablePriceWithinDuration(bytes32 id, uint64 duration) external view returns (PythStructs.Price memory price);
+    function getLatestAvailablePriceWithinDuration(bytes32 id, uint64 duration)
+        external
+        view
+        returns (PythStructs.Price memory price);
 
     /// @notice Update price feeds with given update messages.
     /// This method requires the caller to pay a fee in wei; the required fee can be computed by calling
@@ -137,7 +152,8 @@ interface IPyth {
 
     /// @notice Wrapper around updatePriceFeeds that rejects fast if a price update is not necessary. A price update is
     /// necessary if the current on-chain publishTime is older than the given publishTime. It relies solely on the
-    /// given `publishTimes` for the price feeds and does not read the actual price update publish time within `updateData`.
+    /// given `publishTimes` for the price feeds and
+    /// does not read the actual price update publish time within `updateData`.
     ///
     /// This method requires the caller to pay a fee in wei; the required fee can be computed by calling
     /// `getUpdateFee` with the length of the `updateData` array.
@@ -151,10 +167,14 @@ interface IPyth {
     /// @param updateData Array of price update data.
     /// @param priceIds Array of price ids.
     /// @param publishTimes Array of publishTimes. `publishTimes[i]` corresponds to known `publishTime` of `priceIds[i]`
-    function updatePriceFeedsIfNecessary(bytes[] memory updateData, bytes32[] memory priceIds, uint64[] memory publishTimes) external payable;
+    function updatePriceFeedsIfNecessary(
+        bytes[] memory updateData,
+        bytes32[] memory priceIds,
+        uint64[] memory publishTimes
+    ) external payable;
 
     /// @notice Returns the required fee to update an array of price updates.
     /// @param updateDataSize Number of price updates.
     /// @return feeAmount The required fee in Wei.
-    function getUpdateFee(uint updateDataSize) external view returns (uint feeAmount);
+    function getUpdateFee(uint256 updateDataSize) external view returns (uint256 feeAmount);
 }
