+++ contracts/test/MockPyth.sol
@@ -14,13 +14,16 @@ contract MockPyth is IPyth {
     // simply update price feeds
     function updatePriceFeedsHarness(PythStructs.PriceFeed[] calldata feeds) external {
         require(feeds.length > 0, "feeds length must > 0");
-        for (uint i = 0; i < feeds.length; ++i) {
+        for (uint256 i = 0; i < feeds.length; ++i) {
             priceFeeds[feeds[i].id] = feeds[i];
         }
     }
 
     // a very simple version get price with timestamp expiration check, just for test
-    function getLatestAvailablePriceWithinDuration(bytes32 id, uint64 duration) external view override
+    function getLatestAvailablePriceWithinDuration(bytes32 id, uint64 duration)
+        external
+        view
+        override
         returns (PythStructs.Price memory price)
     {
         PythStructs.PriceFeed memory priceFeed = queryPriceFeed(id);
@@ -35,7 +38,7 @@ contract MockPyth is IPyth {
         return price;
     }
 
-    function _diff(uint x, uint y) internal pure returns (uint) {
+    function _diff(uint256 x, uint256 y) internal pure returns (uint256) {
         if (x > y) {
             return x - y;
         } else {
@@ -45,34 +48,41 @@ contract MockPyth is IPyth {
 
     // not implemented
     function updatePriceFeedsIfNecessary(
-        bytes[] memory updateData, 
-        bytes32[] memory priceIds, 
-        uint64[] memory publishTimes) external payable 
-    {
+        bytes[] memory updateData,
+        bytes32[] memory priceIds,
+        uint64[] memory publishTimes
+    ) external payable {
         updateData;
         priceIds;
         publishTimes;
     }
-    function getUpdateFee(uint updateDataSize) external pure override returns (uint feeAmount) {
+
+    function getUpdateFee(uint256 updateDataSize) external pure override returns (uint256 feeAmount) {
         updateDataSize;
         return 0;
     }
+
     function updatePriceFeeds(bytes[] memory updateData) external payable {
         updateData;
     }
+
     function getCurrentPrice(bytes32 id) external pure override returns (PythStructs.Price memory price) {
         id;
-        return PythStructs.Price({ price: 0, conf: 0, expo: 0});
+        return PythStructs.Price({ price: 0, conf: 0, expo: 0 });
     }
+
     function getEmaPrice(bytes32 id) external pure override returns (PythStructs.Price memory price) {
         id;
-        return PythStructs.Price({ price: 0, conf: 0, expo: 0});
+        return PythStructs.Price({ price: 0, conf: 0, expo: 0 });
     }
-    function getLatestAvailablePriceUnsafe(bytes32 id) external pure override
+
+    function getLatestAvailablePriceUnsafe(bytes32 id)
+        external
+        pure
+        override
         returns (PythStructs.Price memory price, uint64 publishTime)
     {
         id;
-        return (PythStructs.Price({ price: 0, conf: 0, expo: 0}), 0);
+        return (PythStructs.Price({ price: 0, conf: 0, expo: 0 }), 0);
     }
-
 }
